backend/requirements.txt
# FastAPI & Server
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.25
alembic==1.13.1

# Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.1.2

# Data Processing
pandas==2.1.4
openpyxl==3.1.2
xlrd==2.0.1
xlsxwriter==3.1.9
python-dateutil==2.8.2

# Validation
pydantic==2.5.3
pydantic-settings==2.1.0
email-validator==2.1.0

# CORS
starlette==0.35.1

# Utilities
python-dotenv==1.0.0
aiofiles==23.2.1

